# src/docs/openapi2/swagger.yaml
definitions:
...
  consts.ResourceName:
    enum:
    - system
    - audit
    - container
    - container_version
    - dataset
    - dataset_version
    - project
    - label
    - user
    - role
    - permission
    - task
    - trace
    - injection
    - execution
    type: string
    x-enum-comments:
      ResourceAudit: audit resource
      ResourceContainer: container resource
      ResourceContainerVersion: container version resource
      ResourceDataset: dataset resource
      ResourceDatasetVersion: dataset version resource
      ResourceExecution: execution resource
      ResourceInjection: fault injection resource
      ResourceLabel: label resource
      ResourcePermission: permission resource
      ResourceProject: project resource
      ResourceRole: role resource
      ResourceSystem: system resource
      ResourceTask: task resource
      ResourceTrace: trace resource
      ResourceUser: user resource
    x-enum-varnames:
    - ResourceSystem
    - ResourceAudit
    - ResourceContainer
    - ResourceContainerVersion
    - ResourceDataset
    - ResourceDatasetVersion
    - ResourceProject
    - ResourceLabel
    - ResourceUser
    - ResourceRole
    - ResourcePermission
    - ResourceTask
    - ResourceTrace
    - ResourceInjection
    - ResourceExecution
...
  dto.BatchEvaluateDatapackReq:
    properties:
      specs:
        items:
          $ref: '#/definitions/dto.EvaluateDatapackSpec'
        type: array
    required:
    - specs
    type: object
  dto.BatchEvaluateDatapackResp:
    properties:
      failed_count:
        type: integer
      failed_items:
        items:
          type: string
        type: array
      success_count:
        type: integer
      success_items:
        items:
          $ref: '#/definitions/dto.EvaluateDatapackItem'
        type: array
    type: object
  dto.BatchEvaluateDatasetResp:
    properties:
      failed_count:
        type: integer
      failed_items:
        items:
          type: string
        type: array
      success_count:
        type: integer
      success_items:
        items:
          $ref: '#/definitions/dto.EvaluateDatasetItem'
        type: array
    type: object
...
  dto.ChangePasswordReq:
    properties:
      new_password:
        example: newpassword123
        minLength: 8
        type: string
      old_password:
        example: oldpassword123
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  dto.ContainerRef:
    properties:
      name:
        type: string
      version:
        type: string
    required:
    - name
    type: object
...
  dto.DetectorResultItem:
    properties:
      abnormal_avg_duration:
        type: number
      abnormal_p90:
        type: number
      abnormal_p95:
        type: number
      abnormal_p99:
        type: number
      abnormal_succ_rate:
        type: number
      issues:
        type: string
      normal_avg_duration:
        type: number
      normal_p90:
        type: number
      normal_p95:
        type: number
      normal_p99:
        type: number
      normal_succ_rate:
        type: number
      span_name:
        type: string
    required:
    - issues
    - span_name
    type: object
  dto.EvaluateDatapackItem:
    properties:
      algorithm:
        type: string
      algorithm_version:
        type: string
      datapack:
        type: string
      execution_refs:
        items:
          $ref: '#/definitions/dto.ExecutionGranularityRef'
        type: array
      groundtruth:
        $ref: '#/definitions/handler.Groundtruth'
    type: object
  dto.EvaluateDatapackRef:
    properties:
      datapack:
        type: string
      detector_results:
        description: Detector results
        items:
          $ref: '#/definitions/dto.DetectorResultItem'
        type: array
      executed_at:
        description: Execution time
        type: string
      execution_duration:
        description: Execution duration in seconds
        type: number
      execution_id:
        description: Execution ID
        type: integer
      predictions:
        description: Algorithm predictions
        items:
          $ref: '#/definitions/dto.GranularityResultItem'
        type: array
    type: object
  dto.EvaluateDatapackSpec:
    properties:
      algorithm:
        $ref: '#/definitions/dto.ContainerRef'
      datapack:
        type: string
      filter_labels:
        items:
          $ref: '#/definitions/dto.LabelItem'
        type: array
    required:
    - algorithm
    - datapack
    type: object
  dto.EvaluateDatasetItem:
    properties:
      algorithm:
        description: Algorithm name
        type: string
      algorithm_version:
        description: Algorithm version
        type: string
      dataset:
        description: Dataset name
        type: string
      dataset_version:
        description: Dataset version
        type: string
      evalaute_refs:
        description: Evaluation refs for each dataset
        items:
          $ref: '#/definitions/dto.EvaluateDatapackRef'
        type: array
      executed_count:
        description: Number of successfully executed datapacks
        type: integer
      total_count:
        description: Total number of datapacks in dataset
        type: integer
    type: object
...
  dto.ExecutionGranularityRef:
    properties:
      detector_results:
        description: Detector results
        items:
          $ref: '#/definitions/dto.DetectorResultItem'
        type: array
      executed_at:
        description: Execution time
        type: string
      execution_duration:
        description: Execution duration in seconds
        type: number
      execution_id:
        description: Execution ID
        type: integer
      predictions:
        description: Algorithm predictions
        items:
          $ref: '#/definitions/dto.GranularityResultItem'
        type: array
    type: object
...
  dto.GenericResponse-any:
    properties:
      code:
        description: Status code
        type: integer
      data:
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_BatchEvaluateDatapackResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.BatchEvaluateDatapackResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
  dto.GenericResponse-dto_BatchEvaluateDatasetResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.BatchEvaluateDatasetResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_LoginResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.LoginResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_TokenRefreshResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.TokenRefreshResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_UserDetailResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.UserDetailResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
  dto.GenericResponse-dto_UserInfo:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.UserInfo'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GranularityResultItem:
    properties:
      confidence:
        type: number
      level:
        type: string
      rank:
        type: integer
      result:
        type: string
    required:
    - level
    - rank
    - result
    type: object
...
  dto.LabelItem:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  dto.LoginReq:
    properties:
      password:
        example: password123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResp:
    properties:
      expires_at:
        example: "2024-12-31T23:59:59Z"
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/dto.UserInfo'
    type: object
...
  dto.PermissionResp:
    properties:
      action:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      resource_name:
        $ref: '#/definitions/consts.ResourceName'
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.RegisterReq:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      username:
        example: newuser
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  dto.RoleResp:
    properties:
      display_name:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.TokenRefreshReq:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - token
    type: object
  dto.TokenRefreshResp:
    properties:
      expires_at:
        example: "2024-12-31T23:59:59Z"
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  dto.UserContainerInfo:
    properties:
      container_id:
        type: integer
      container_name:
        type: string
      joined_at:
        type: string
      role_name:
        type: string
    type: object
  dto.UserDatasetInfo:
    properties:
      dataset_id:
        type: integer
      dataset_name:
        type: string
      joined_at:
        type: string
      role_name:
        type: string
    type: object
  dto.UserDetailResp:
    properties:
      avatar:
        type: string
      container_roles:
        items:
          $ref: '#/definitions/dto.UserContainerInfo'
        type: array
      created_at:
        type: string
      dataset_roles:
        items:
          $ref: '#/definitions/dto.UserDatasetInfo'
        type: array
      email:
        type: string
      full_name:
        type: string
      global_roles:
        items:
          $ref: '#/definitions/dto.RoleResp'
        type: array
      id:
        type: integer
      is_active:
        type: boolean
      last_login_at:
        type: string
      permissions:
        items:
          $ref: '#/definitions/dto.PermissionResp'
        type: array
      phone:
        type: string
      project_roles:
        items:
          $ref: '#/definitions/dto.UserProjectInfo'
        type: array
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  dto.UserInfo:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      id:
        example: 1
        type: integer
      username:
        example: admin
        type: string
    type: object
  dto.UserProjectInfo:
    properties:
      joined_at:
        type: string
      project_id:
        type: integer
      project_name:
        type: string
      role_name:
        type: string
    type: object
...
  handler.Groundtruth:
    properties:
      container:
        items:
          type: string
        type: array
      function:
        items:
          type: string
        type: array
      metric:
        items:
          type: string
        type: array
      pod:
        items:
          type: string
        type: array
      service:
        items:
          type: string
        type: array
      span:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8082
info:
  contact:
    email: team@rcabench.com
    name: RCABench Team
  description: |-
    RCABench - A comprehensive root cause analysis benchmarking platform for microservices
    This API provides endpoints for managing datasets, algorithms, evaluations, and fault injections
    for root cause analysis in distributed systems and microservices architectures.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: RCABench API
  version: 1.0.1
paths:
...
  /api/v2/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change password for authenticated user
      operationId: change_password
      parameters:
      - description: Password change request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "400":
          description: Invalid request format/parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - Authentication
  /api/v2/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password
      operationId: login
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_LoginResp'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Invalid user name or password
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      summary: User login
      tags:
      - Authentication
  /api/v2/auth/logout:
    post:
      description: Logout user and invalidate token
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "400":
          description: Invalid authorization header
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Invalid token
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      summary: User logout
      tags:
      - Authentication
  /api/v2/auth/profile:
    get:
      description: Get profile information for authenticated user
      operationId: get_current_user_profile
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_UserDetailResp'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Authentication
  /api/v2/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh an existing JWT token
      operationId: refresh_auth_token
      parameters:
      - description: Token refresh request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TokenRefreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_TokenRefreshResp'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Invalid token
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      summary: Refresh JWT token
      tags:
      - Authentication
  /api/v2/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      operationId: register_user
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_UserInfo'
        "400":
          description: Invalid request format/parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      summary: User registration
      tags:
      - Authentication
...
  /api/v2/evaluations/datapacks:
    post:
      consumes:
      - application/json
      description: Retrieve evaluation data for multiple algorithm-datapack pairs.
      operationId: evaluate_algorithm_on_datapacks
      parameters:
      - description: Batch evaluation request containing multiple algorithm-datapack
          pairs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchEvaluateDatapackReq'
      produces:
      - application/json
      responses:
        "200":
          description: Batch algorithm datapack evaluation data retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_BatchEvaluateDatapackResp'
        "400":
          description: Invalid request format/parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: List Datapack Evaluation Results
      tags:
      - Evaluations
  /api/v2/evaluations/datasets:
    post:
      consumes:
      - application/json
      description: Retrieve evaluation data for multiple algorithm-dataset pairs.
      operationId: evaluate_algorithm_on_datasets
      parameters:
      - description: Batch evaluation request containing multiple algorithm-dataset
          pairs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.BatchEvaluateDatapackReq'
      produces:
      - application/json
      responses:
        "200":
          description: Batch algorithm dataset evaluation data retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_BatchEvaluateDatasetResp'
        "400":
          description: Invalid request format/parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: List Dataset Evaluation Results
      tags:
      - Evaluations
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"